requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
pypdf
pydantic
python-dotenv

sentence-transformers